Refactor: 알람 api 서비스

+++ src/main/java/com/openbook/openbook/user/service/AlarmService.java
@@ -1,8 +1,9 @@
-package com.openbook.openbook.user.service.core;
+package com.openbook.openbook.user.service;
 
 
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
+import com.openbook.openbook.user.dto.AlarmDto;
 import com.openbook.openbook.user.entity.dto.AlarmType;
 import com.openbook.openbook.user.entity.Alarm;
 import com.openbook.openbook.user.entity.User;
@@ -11,16 +12,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class AlarmService {
 
+    private final UserService userService;
     private final AlarmRepository alarmRepository;
-
-    public boolean userHasReceivedAlarms(Long receiverId) {
-        return alarmRepository.existsByReceiverId(receiverId);
-    }
 
     public Alarm getAlarmOrException(Long id) {
         return alarmRepository.findById(id).orElseThrow(()->
@@ -40,15 +39,32 @@ public class AlarmService {
         );
     }
 
-    public Slice<Alarm> getUserReceivedAlarm(Pageable pageable, User receiver) {
-        return alarmRepository.findAllByReceiver(pageable, receiver);
+    @Transactional(readOnly = true)
+    public Slice<AlarmDto> getAlarmData(Pageable pageable, final Long id) {
+        User user = userService.getUserOrException(id);
+        return alarmRepository.findAllByReceiver(pageable, user).map(AlarmDto::of);
     }
 
-    public void deleteAlarm(Alarm alarm) {
+    @Transactional
+    public void deleteAlarm(final Long userId, final Long alarmId) {
+        Alarm alarm = getAlarmOrException(alarmId);
+        if(alarm.getReceiver().getId()!=userId) {
+            throw new OpenBookException(ErrorCode.FORBIDDEN_ACCESS);
+        }
         alarmRepository.delete(alarm);
     }
 
-    public void deleteAllReceiverAlarm(Long receiverId) {
-        alarmRepository.deleteAllByReceiverId(receiverId);
+    @Transactional
+    public void deleteAllAlarm(final Long userId) {
+        User user = userService.getUserOrException(userId);
+        if(!userHasReceivedAlarms(userId)) {
+            throw new OpenBookException(ErrorCode.ALARM_NOT_FOUND);
+        }
+        alarmRepository.deleteAllByReceiverId(user.getId());
     }
+
+    private boolean userHasReceivedAlarms(Long receiverId) {
+        return alarmRepository.existsByReceiverId(receiverId);
+    }
+
 }
